Create a matrix with all data for the Data2 matrix.

Return a matrix with data almost ready to be cut and pasted into the main data file (Data2_1group.odc) for the model.

format_Data2(fish, env, ndays, missing_days = NULL)

Arguments

fish

data frame with three mandatory columns "marked", "capture_day" and "recapture_day"

env

data frame with two mandatory columns "w_temp" and "w_level".

ndays

numeric number of days data have in operation, e.g. start date to stop date

missing_days

numeric vector of day numbers when capture (and recapture) didn't work.

Value

a matrix with ndays rows and ndays + 6 columns

Details

Each row data frame fish (with mandatory columns "marked", "capture_day" and "recapture_day") represent a caught fish. Column capture_day contains the number of days between the start of the experiment the day the fish was caught for the first time. Column recapture_day is the same if recaptured and NA otherwise. Column marked is set to TRUE for all fish that was marked and subject for for recapture and FALSE otherwise, e.g. when marked is false the fish was not marked and not subject for possible recapture.

Data frame env must have ndays rows and two mandatory columns w_temp (water temperature) and w_level (water level). No NAs allowed so impute the data if you have missing values.

Ndays is the number of days the experiment was running including any downtime. Optional parameter missing_dats is a vector with day numbers when the experiment was put on hiatus.

Examples

ndays <- 5
(smolts <- data.frame(
  capture_day = c(2,2,3),
  recapture_day = c(NA, 3, 4),
  marked = c(FALSE, TRUE, TRUE)))
#>   capture_day recapture_day marked
#> 1           2            NA  FALSE
#> 2           2             3   TRUE
#> 3           3             4   TRUE
(envdata <- data.frame(
  w_temp = c(7, 7, 8, 8.5, 9),
  w_level = c(100, 110, 105, 150, 145)))
#>   w_temp w_level
#> 1    7.0     100
#> 2    7.0     110
#> 3    8.0     105
#> 4    8.5     150
#> 5    9.0     145

format_Data2(fish = smolts, env = envdata, ndays = ndays)
#>      m[,1] c[,1] r[,1,1] r[,2,1] r[,3,1] r[,4,1] r[,5,1] wt[,1] wl[,1] z[,1]
#> [1,]     0     0       0       0       0       0       0    7.0    100     0
#> [2,]     1     2       0       0       1       0       0    7.0    110     0
#> [3,]     1     1       0       0       0       1       0    8.0    105     0
#> [4,]     0     0       0       0       0       0       0    8.5    150     0
#> [5,]     0     0       0       0       0       0       0    9.0    145     0
#>      n[,1]
#> [1,]     0
#> [2,]     0
#> [3,]     0
#> [4,]     0
#> [5,]     0